requirements.txt
# SEO Causal Test Simulator - Project Dependencies
# Python >= 3.8

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Causal Inference
pycausalimpact>=0.1.1
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Visualization & Reporting
plotly>=5.17.0
matplotlib>=3.7.0
streamlit>=1.28.0
reportlab>=4.0.0

# Data Storage & Configuration
duckdb>=0.8.0
pyyaml>=6.0
python-dateutil>=2.8.0
